// types.ts
export interface SlideProps {
  isActive?: boolean;
  isExporting?: boolean;
}

export interface StudentData {
  faculty: string;
  count: number;
}

export interface WaterData {
  month: string;
  value: number;
}

export interface FrequencyData {
  value: number;
  freq: number;
}

export interface IntervalData {
  interval: string;
  freq: number;
}